Add files via upload
This is my code from the DengueAI DrivenData competition where we were asked to predict the spread of Dengue Fever. I scored in the top 25% of this competition, and it was a fun opportunity to further my understanding of stacking algorithms and figure out how to find patterns within data. Another great part of this competition for me was practicing my skills with imputing based on other features.